Separated MyWsHandler from the main
